buildout/production-update/requirements.txt
Fabric==1.8.0
path.py==5.0
pycrypto==2.6.1
ssh==1.8.0
